Fix null String that caused catastrophic failure
The problem was with what I changed in the switch statement. I saw that we were calling the SmartDashboard in most cases, and I thought, "Why not set a variable in each case and just reference the SmartDashboard once?" With this goal, I declared a `String`, which Eclipse wanted me to initialize. I didn't know what to do, because I'd read that Strings are immutable in Java, which in my mind meant that once you defined a String, you couldn't change it. I took a wild, foolish guess and set it equal to `null`, since it seemed reasonable to start a string out empty. I still don't know what `null` means in this context—only that it was the wrong choice. Once we reverted the changes in this file, the code worked, as though by magic. The clue was the green message in the Driver Station console, which indicated the source of the failure.